update to snapshot spandsp-20090421
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@13086 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/spandsp/tests/fsk_tests.c
@@ -22,7 +22,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: fsk_tests.c,v 1.55 2009/02/12 14:21:16 steveu Exp $
+ * $Id: fsk_tests.c,v 1.56 2009/04/14 16:04:54 steveu Exp $
  */
 
 /*! \page fsk_tests_page FSK modem tests
@@ -392,7 +392,7 @@ int main(int argc, char *argv[])
         bert_set_report(&caller_bert, 100000, reporter, (void *) (intptr_t) 1);
         bert_init(&answerer_bert, bits_per_test, BERT_PATTERN_ITU_O152_11, test_bps, 20);
         bert_set_report(&answerer_bert, 100000, reporter, (void *) (intptr_t) 2);
-        if ((model = both_ways_line_model_init(line_model_no, (float) noise_level, line_model_no, noise_level, channel_codec, rbs_pattern)) == NULL)
+        if ((model = both_ways_line_model_init(line_model_no, (float) noise_level, line_model_no, (float) noise_level, channel_codec, rbs_pattern)) == NULL)
         {
             fprintf(stderr, "    Failed to create line model\n");
             exit(2);
